Separate ServiceConfigDisplay sections
