+++ plugins/services/src/js/service-configuration/ServiceConfigDisplay.js
@@ -6,8 +6,23 @@ import ConfigurationMapLabel from '../../../../../src/js/components/Configuratio
 import ConfigurationMapRow from '../../../../../src/js/components/ConfigurationMapRow';
 import ConfigurationMapSection from '../../../../../src/js/components/ConfigurationMapSection';
 import ConfigurationMapValue from '../../../../../src/js/components/ConfigurationMapValue';
+import GeneralServiceConfigSection from './GeneralServiceConfigSection';
+import NetworkingServiceConfigSection from './NetworkingServiceConfigSection';
+import StorageServiceConfigSection from './StorageServiceConfigSection';
+import EnvironmentVariablesServiceConfigSection from './EnvironmentVariablesServiceConfigSection';
+import LabelsServiceConfigSection from './LabelsServiceConfigSection';
+import HealthChecksServiceConfigSection from './HealthChecksServiceConfigSection';
 import ServiceConfigDisplayUtil from '../utils/ServiceConfigDisplayUtil';
 import Util from '../../../../../src/js/utils/Util';
+
+const serviceConfigDisplayList = [
+  GeneralServiceConfigSection,
+  NetworkingServiceConfigSection,
+  StorageServiceConfigSection,
+  EnvironmentVariablesServiceConfigSection,
+  LabelsServiceConfigSection,
+  HealthChecksServiceConfigSection
+];
 
 const shouldExcludeItem = (row, appConfig) => {
   switch (row.key) {
@@ -44,45 +59,45 @@ const getDisplayValue = (type, value) => {
 };
 
 const ServiceConfigDisplay = ({appConfig}) => {
-  const configurationMapSections = ServiceConfigDisplayUtil.displayValues.map(
+  const configurationMapSections = serviceConfigDisplayList.map(
     (section, sectionIndex) => {
       let configurationMapRows = section.values.filter((row) => {
-          // Some rows must be excluded if relevant data is missing.
-          return !shouldExcludeItem(row, appConfig);
-        })
-        .map((row, rowIndex) => {
-          let reactKey = `${sectionIndex}.${rowIndex}`;
-          let value = Util.findNestedPropertyInObject(appConfig, row.key);
+        // Some rows must be excluded if relevant data is missing.
+        return !shouldExcludeItem(row, appConfig);
+      })
+      .map((row, rowIndex) => {
+        let reactKey = `${sectionIndex}.${rowIndex}`;
+        let value = Util.findNestedPropertyInObject(appConfig, row.key);
 
-          // If a transformValue was specified on the row, we use it.
-          if (row.transformValue != null) {
-            value = row.transformValue(value, appConfig);
-          }
+        // If a transformValue was specified on the row, we use it.
+        if (row.transformValue != null) {
+          value = row.transformValue(value, appConfig);
+        }
 
-          if (row.render != null) {
-            // If a custom render method was specified on the row, we use that.
-            return row.render(value, appConfig);
-          } else if (row.heading != null) {
-            // If the row is a heading, we render the heading.
-            return (
-              <ConfigurationMapHeading key={reactKey} level={row.headingLevel}>
-                {row.heading}
-              </ConfigurationMapHeading>
-            );
-          }
-
-          // Otherwise we treat the row as "label:value" type display.
+        if (row.render != null) {
+          // If a custom render method was specified on the row, we use that.
+          return row.render(value, appConfig);
+        } else if (row.heading != null) {
+          // If the row is a heading, we render the heading.
           return (
-            <ConfigurationMapRow key={reactKey}>
-              <ConfigurationMapLabel>
-                {row.label}
-              </ConfigurationMapLabel>
-              <ConfigurationMapValue>
-                {getDisplayValue(row.type, value)}
-              </ConfigurationMapValue>
-            </ConfigurationMapRow>
+            <ConfigurationMapHeading key={reactKey} level={row.headingLevel}>
+              {row.heading}
+            </ConfigurationMapHeading>
           );
-        });
+        }
+
+        // Otherwise we treat the row as "label:value" type display.
+        return (
+          <ConfigurationMapRow key={reactKey}>
+            <ConfigurationMapLabel>
+              {row.label}
+            </ConfigurationMapLabel>
+            <ConfigurationMapValue>
+              {getDisplayValue(row.type, value)}
+            </ConfigurationMapValue>
+          </ConfigurationMapRow>
+        );
+      });
 
       if (configurationMapRows.length === 0) {
         return null;
